Add option for username to headless mode
Support a username and password when running in headless mode from
a config file. This tests the normal installer functionality of setting a
username other than 'pentoo'.

Used by the ISO tester tool to set the username to 'vagrant'.
Vagrant software is used to run the installed pentoo with 'vagrant up'
and I can't get 'vagrant up' to work with any other username.