Enable post installation commands from config file
Used by the ISO tester tool to place a ssh key to
~/.ssh/authorized_keys so 'vagrant up' can run the installed pentoo.

The advantage of the solution in this commit is that everything is mounted
already in the installer. And being able to simply run such
parameterized scripts in the installer might come handy for other
purposes.
The disadvantage is that this feature of the installer is not needed for
normal installer operation.

An alternative would be to do it in the ISO tester tool by mounting the
root partition after running the installer.
While this alternative works well for simple disk layouts, it would
be hard to handle when user home is not on root partition or when
partitions are setup encrypted by the installer.